+++ simulator.py
@@ -9,7 +9,7 @@ import csv
 import os
 from logging_config import setup_logging, log_exceptions, get_logger
 import logging
-from SolarCoreSim_v_3_5 import generate_report_off_grid, plot_energy_data, plot_battery_profile
+from visualization import generate_report_off_grid, plot_energy_data, plot_battery_profile
 from config import load_config
 
 class Simulator:
